backend/web_api/requirements.txt
django==1.11
djangorestframework==3.10
django-cors-headers==2.2.0